split files (one per type)
