+++ lib/metrics/AmqpTopologyBuilder.ts
@@ -1,0 +1,8 @@
+import "source-map-support/register";
+
+import { Exchange, Queue } from "amqp-ts";
+
+/**
+ * Interface for building the underlying amqp topology.
+ */
+export type AmqpTopologyBuilder = () => Queue | Exchange;
