+++ lib/metrics/ICounterValue.ts
@@ -1,0 +1,8 @@
+import "source-map-support/register";
+
+/**
+ * Interface used when extracting values from a {@link Counter} or {@link MonotoneCounter}.
+ */
+export interface ICounterValue {
+  count: number;
+}
