+++ lib/metrics/index.ts
@@ -1,3 +1,12 @@
 import "source-map-support/register";
 
 export * from "./AmqpMetricReporter";
+export * from "./AmqpMetricReporterOptions";
+export * from "./AmqpTopologyBuilder";
+export * from "./AmqpTopologyHelper";
+export * from "./ICounterValue";
+export * from "./IGaugeValue";
+export * from "./IHistogramValue";
+export * from "./IMeterValue";
+export * from "./ITimerValue";
+export * from "./MetricMessageBuilder";
